refactor: ListView를 RecyclerView로 수정
