+++ app/src/main/java/woowacourse/movie/ui/movielist/MovieListAdapter.kt
@@ -3,68 +3,23 @@ package woowacourse.movie.ui.movielist
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.BaseAdapter
-import android.widget.Button
-import android.widget.ImageView
-import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.model.MovieModel
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
 
 class MovieListAdapter(
     private val movies: List<MovieModel>,
-    private val itemButtonClickListener: ItemButtonClickListener,
-) : BaseAdapter() {
-    interface ItemButtonClickListener {
-        fun onClick(position: Int)
+    private val onItemClick: (Int) -> Unit
+) : RecyclerView.Adapter<MovieListViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieListViewHolder {
+        val view: View =
+            LayoutInflater.from(parent.context).inflate(R.layout.movie_item, parent, false)
+        return MovieListViewHolder(view, onItemClick)
     }
 
-    override fun getCount(): Int = movies.size
+    override fun getItemCount(): Int = movies.size
 
-    override fun getItem(position: Int): Any = movies[position]
-
-    override fun getItemId(position: Int): Long = position.toLong()
-
-    override fun getView(position: Int, convertView: View?, parent: ViewGroup?): View {
-        val movie = movies[position]
-        val viewHolder: MovieListViewHolder
-
-        val itemView = if (convertView != null) {
-            convertView.also { viewHolder = it.tag as MovieListViewHolder }
-        } else {
-            val view = LayoutInflater.from(parent?.context).inflate(R.layout.movie_item, parent, false)
-            viewHolder = MovieListViewHolder(view)
-            view.tag = viewHolder
-            view
-        }
-
-        viewHolder.bind(movie)
-        viewHolder.setButtonClickListener(position)
-
-        return itemView
-    }
-
-    inner class MovieListViewHolder(
-        private val view: View,
-    ) {
-        private val poster: ImageView = view.findViewById(R.id.item_poster)
-        private val title: TextView = view.findViewById(R.id.item_title)
-        private val date: TextView = view.findViewById(R.id.item_date)
-        private val runningTime: TextView = view.findViewById(R.id.item_running_time)
-        private val button: Button = view.findViewById(R.id.item_booking_button)
-
-        fun setButtonClickListener(position: Int) {
-            button.setOnClickListener { itemButtonClickListener.onClick(position) }
-        }
-
-        fun bind(movie: MovieModel) {
-            poster.setImageResource(movie.poster)
-            title.text = movie.title
-            date.text = view.context.getString(R.string.screening_date, movie.startDate.format(), movie.endDate.format())
-            runningTime.text = view.context.getString(R.string.running_time, movie.runningTime)
-        }
-
-        private fun LocalDate.format(): String = format(DateTimeFormatter.ofPattern(view.context.getString(R.string.date_format)))
+    override fun onBindViewHolder(holder: MovieListViewHolder, position: Int) {
+        holder.bind(movies[position])
     }
 }
