+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     tools:context=".ui.movielist.MainActivity">
 
-    <ListView
+    <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/main_movie_list"
         android:layout_width="0dp"
         android:layout_height="0dp"
